+++ RELEASE_NOTES.rst
@@ -33,6 +33,7 @@ Improvements
   thanks to Miroslav Šedivý).
 - Several improvements to wheels generation in CI
   (thanks to Andreas Motl @amotl and Matthias @xmatthias).
+- Simplified management of version information.
 
 Bugfixes
 --------
@@ -44,3 +45,6 @@ Bugfixes
 Other changes
 -------------
 - Drop the deprecated `hdf5Version` and `File.open_count`.
+- the :func:`get_tables_version` and :func:`get_hdf5_version` functions are
+  now deprecated please use the coresponding :data:`tables.__version__` and
+  :data:`tables.hdf5_version` instead.